ENH: cleanup 'importer.py', and move back 'do_nedd' from 'general_data' --- this method is different.
